requirements.txt
pyserial>=3.5
numpy>=1.21,<1.22
